# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kapital API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Kapital backend services"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/repayments/pending": {
      "get": {
        "summary": "Get pending repayment requests (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending repayment requests"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/repayments/{id}/approve": {
      "post": {
        "summary": "Approve a pending repayment (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet transaction ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notes": {
                    "type": "string",
                    "description": "Admin notes for the approval"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repayment approved successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/repayments/{id}/reject": {
      "post": {
        "summary": "Reject a pending repayment (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet transaction ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for rejection"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional admin notes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repayment rejected successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/admin/loans/{id}/request-discharge": {
      "post": {
        "summary": "Request loan discharge (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The loan ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for requesting discharge"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discharge request submitted successfully"
          },
          "400": {
            "description": "Invalid request or loan cannot be discharged"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/loans/{id}/approve-discharge": {
      "post": {
        "summary": "Approve loan discharge (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The loan ID"
          }
        ],
        "requestBody": {
          "required": false
        },
        "responses": {
          "200": {
            "description": "Loan discharged successfully"
          },
          "400": {
            "description": "Invalid request or loan cannot be discharged"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/admin/loans/{id}/reject-discharge": {
      "post": {
        "summary": "Reject loan discharge request (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The loan ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for rejecting the discharge"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discharge request rejected successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied, admin privileges required"
          },
          "404": {
            "description": "Loan not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}